requirements.txt
streamlit
openai==0.28.0
python-dotenv
selenium==4.15.0
webdriver-manager